Add fromAddress to BroadcastOptions for HyperCore transfer
Introduces a fromAddress field to BroadcastOptions for improved transaction context, updates initializers and mapping extensions, and refactors related code to support the new parameter. Also updates GatewayService to use public extension methods and fixes UTXO mapping to preserve vout type.

+++ Packages/GemstonePrimitives/Sources/Extensions/GemUtxo+GemstonePrimitives.swift
@@ -8,7 +8,7 @@ extension GemUtxo {
     public func map() throws -> UTXO {
         UTXO(
             transaction_id: transactionId,
-            vout: Int32(vout),
+            vout: vout,
             value: value,
             address: address
         )
@@ -19,7 +19,7 @@ extension UTXO {
     public func map() -> GemUtxo {
         GemUtxo(
             transactionId: transaction_id,
-            vout: UInt32(vout),
+            vout: vout,
             value: value,
             address: address
         )
